refactor preprocessing logic from topic module into module of its own
